Refactor guard test utils
